--- advent_of_code/aoc/2022/inputs/day22_sample.txt
@@ -1,14 +1,0 @@
-        ...#
-        .#..
-        #...
-        ....
-...#.......#
-........#...
-..#....#....
-..........#.
-        ...#....
-        .....#..
-        .#......
-        ......#.
-
-10R5L5R10L4R5L5
